support reference bundle in build-on-commit action
Add an optional field so users can specify the reference bundle to use when
testing their charm. If set, the action will build the charm locally and
deploy it. It will then deploy the given reference bundle. Bundletester
will re-use the deployed local charm in place of the one specified in
bundle.yaml.