Work around intermittent install hook failure
Explicitly add PyYAML==3.11 to wheelhouse.txt because, depending on the
order that the wheels are collected during the install hook,
jujubundlelib may fail without it in the wheelhouse.  Actually, the real
issue is that it incorrectly passes as long as charmhelpers gets picked
up from the wheelhouse first, because that accepts PyYAML 3.12 from the
wheelhouse and jujubundlelib's exact 3.11 requirement gets ignored.
That's a bug in pip.  :/